Move EditXrefTool to shared folder.
