+++ src/article/shared/EditXrefTool.js
